Revert "moving gem_function to TC_database"
This reverts commit f2a62464bb88019af2f89c8504bc8a2faa3b041a.
