+++ src/dump_function.c
@@ -30,7 +30,7 @@
 #include <unistd.h>
 #include "mpi.h"
 #include "MAGEMin.h"
-#include "all_endmembers.h"
+#include "gem_function.h"
 #include "all_solution_phases.h"
 #include "toolkit.h"
