+++ src/gem_function.h
