+++ src/pp_min_function.c
@@ -23,7 +23,8 @@ This function simply update the driving forces of pure phase each time the G-hyp
 #include <complex.h> 
 
 #include "MAGEMin.h"
-#include "all_endmembers.h"
+#include "gem_function.h"
+
 
 /**
   main pure phase minimization routine
